Non-functional updates (code cleaning).
	modified:   idas.chox.core/src/main/java/idas/chox/core/model/ClaimType.java
	modified:   idas.chox.core/src/main/java/idas/chox/core/model/Insurer.java
	modified:   idas.chox.data/src/main/java/idas/chox/data/services/AuditTrailServiceImpl.java
	modified:   idas.chox.data/src/main/java/idas/chox/data/services/WebUserUserRoleServiceImpl.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/admin/AdminUserService.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/xml/UploadClaimXMLServiceImpl.java
	modified:   idas.chox.service/src/test/java/idas/chox/service/LookupItemTest.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/AlertAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/ClaimActivityAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/InsurerBreBandAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/SearchClaimAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/viewdata/TaskViewData.java
	modified:   wsUploadClient/src/main/java/idas/chox/uploadclient/Upload.java

+++ idas.chox.web/src/main/java/idas/chox/web/actions/ClaimActivityAction.java
@@ -116,8 +116,8 @@ public class ClaimActivityAction extends BaseAction implements ModelDriven<Activ
                             claim.setBreBand(choBand);
                         }
 
-                        if ((getIsInsurer() && claim.getInsurer().getId().intValue() != getAuthenticatedUser().getInsurer().getId().intValue())
-                                || (getIsCHO() && claim.getChorganisation().getId().intValue() != getAuthenticatedUser().getChorganisation().getId().intValue())) {
+                        if ((getIsInsurer() && claim.getInsurer().getId() != getAuthenticatedUser().getInsurer().getId().intValue())
+                                || (getIsCHO() && claim.getChorganisation().getId() != getAuthenticatedUser().getChorganisation().getId().intValue())) {
                             throw new AccessDeniedException("Attempt to access a claim that you do not own.");
                         }
                         activity.process(claim);
